Responses are now continuous.
In stead of workign with the repsonse object, you can now inject the
ChatRoom instance into a method. That allows you to keep sending
messages in your thread as long as you like. This also gets rid of all
the response types etc, makign it all nice and clean.

Because this breaks the API, I bumped up the version to 0.2.0

Closes #10